refactor(settings): extract avatar upload flow into useAvatarUpload
Move avatarUploading state, the file-input ref, and the upload/remove
handlers verbatim into _hooks/use-avatar-upload.ts. Hook pulls
uploadAvatar/removeAvatar from useAuth and owns its Settings
translations; result reporting delegated via an onMessage callback so
the page keeps its single message system. Drops the now-unused
uploadAvatar/removeAvatar from the page's useAuth destructure.
Behavior-identical; page.tsx 1547->1522 LOC.
